feat: reorganize SourceControl API with dedicated clients folder
